first

+++ src/Opdracht3.java
@@ -5,9 +5,9 @@ public class Opdracht3 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("Side 1: ");
-        double side1 = scanner.nextInt();
+        double side1 = scanner.nextDouble();
         System.out.print("Side 2: ");
-        double side2 = scanner.nextInt();
+        double side2 = scanner.nextDouble();
         System.out.println(Math.sqrt(side1*side1+side2*side2));
 
     }
